requirements.txt
telegram
configparser
redis
python-telegram-bot==13.7
steamship
requests
openai
mysql-connector-python